Add filter for displaying session info in view.
An action filter was added that modifies the ViewData on a controller in
order to turn "session information" display on or off depending on the
user's currently logged-in status.

The session information currently only includes the user name, but may
be updated to include the number of items in cart, and appropriate login
or checkout buttons.

A new layout was also create to accommodate this, however the
implementation is simply a proof of concept.